Refactored plus docs update
