+++ documentation/docs/examples/click-reactions-pointcloud.mdx
@@ -30,16 +30,3 @@ viewer.on('click', async event => {
   };
 });
 ```
-
-## Controlling the intersection threshold
-
-By default, points within 5 cm of an intersection ray is considered to be intersecting. For point clouds of large
-physical assets this is usually a good value, but for point clouds with large extends (e.g. geographical point clouds)
-this threshold might be too small.
-
-The intersection threshold can be changed using the `pointIntersectionThreshold` option of `getIntersectionFromPixel`:
-
-```js
-const options = { pointIntersectionThreshold: 1.0 /* 1 m */ };
-const intersection = await viewer.getIntersectionFromPixel(event.offsetX, event.offsetY, options);
-```
