+++ examples/src/pages/Migration.tsx
@@ -368,6 +368,7 @@ export function Migration() {
         const start = performance.now();
         const intersection = await viewer.getIntersectionFromPixel(offsetX, offsetY);
         if (intersection !== null) {
+          console.log('intersection', intersection);
           switch (intersection.type) {
             case 'cad':
               {
